+++ app/build.gradle
@@ -12,7 +12,7 @@ android {
     compileSdk 34
 
     defaultConfig {
-        applicationId "com.template.androidbasicapp"
+        applicationId "com.template.androidbasicapp" // TODO applicationIdを他の人と被らない名前に変更してください。
         minSdk 29
         targetSdk 34
         versionCode 100
