+++ app/src/main/java/com/template/androidbasicapp/ui/fragment/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.template.androidbasicapp.ui.gallery;
+package com.template.androidbasicapp.ui.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,23 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.template.androidbasicapp.databinding.FragmentGalleryBinding;
+import com.template.androidbasicapp.databinding.FragmentHomeBinding;
+import com.template.androidbasicapp.ui.viewmodel.HomeViewModel;
 
-public class GalleryFragment extends Fragment {
+public class HomeFragment extends Fragment {
 
-    private FragmentGalleryBinding binding;
+    private FragmentHomeBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        GalleryViewModel galleryViewModel =
-                new ViewModelProvider(this).get(GalleryViewModel.class);
+        HomeViewModel homeViewModel =
+                new ViewModelProvider(this).get(HomeViewModel.class);
 
-        binding = FragmentGalleryBinding.inflate(inflater, container, false);
+        binding = FragmentHomeBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textGallery;
-        galleryViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textHome;
+        homeViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
