+++ app/src/main/res/layout/fragment_slideshow.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.slideshow.SlideshowFragment">
+    tools:context=".ui.fragment.SlideshowFragment">
 
     <TextView
         android:id="@+id/text_slideshow"
