+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.template.androidbasicapp.ui.home.HomeFragment"
+        android:name="com.template.androidbasicapp.ui.fragment.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.template.androidbasicapp.ui.gallery.GalleryFragment"
+        android:name="com.template.androidbasicapp.ui.fragment.GalleryFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.template.androidbasicapp.ui.slideshow.SlideshowFragment"
+        android:name="com.template.androidbasicapp.ui.fragment.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 </navigation>
